product-url-crawler/job-list-api/requirements.txt
google-cloud-firestore>=2.0.0
functions-framework>=3.0.0